feat(invoice): decimal line quantity — accept 1.5 and 1,5, emit + render correctly (S157)
Operator can now invoice fractional units (1.5 consulting days, 0.25
hours). Pre-S157 the quantity was u32 end-to-end: the SPA bound an
integer stepper, the domain LineItem stored u32, and the PDF parser
truncated any decimal NAV quantity to a whole unit.

Wire/storage shape:
- billing LineItem.quantity + issue_invoice::LineJson.quantity +
  print_invoice::ParsedNavLine.quantity + invoice-pdf LineItem.quantity:
  u32 -> rust_decimal::Decimal (exact, no f64 drift).
- net_total = round_half_even(unit_price x quantity) via new
  Huf::checked_mul_decimal (fractional forint/cent cannot exist on wire).
- DuckDB invoice_line.quantity INTEGER -> DECIMAL(18,6). DuckDB forbids
  ALTER COLUMN TYPE on a CHECK-constrained column, so MIGRATE_S157_SQL
  is add/backfill/drop/rename, guarded on the column type to run once.
- Wire JSON quantity is the canonical dot-decimal STRING (C11
  Decimal-as-string convention, as exchange_rate); Decimal serde accepts
  both the string (SPA) and a legacy JSON number (pre-S157 side-store).

SPA: quantity input is text/inputmode=decimal bound to a raw string,
parsed at compose via parseDecimalQuantity (accepts . and , -> same
value, <=0 / >6dp / garbage -> "0" so the backend preflight fires).
formatQuantity renders the Hungarian comma for display.

NAV emit: <quantity> uses Decimal.normalize() (dot-separated, trailing
zeros trimmed). PDF: format::quantity renders the Hungarian comma.
Preflight LineItemQuantityZero now "strictly positive" (HU + EN messages
updated); fractional quantities pass.

+++ apps/aberp/src/issue_invoice.rs
@@ -238,7 +238,13 @@ pub struct CustomerJson {
 #[derive(Debug, Deserialize, Serialize)]
 pub struct LineJson {
     pub description: String,
-    pub quantity: u32,
+    /// S157 — decimal line quantity (1.5 days, 0.25 hours). `Decimal`'s
+    /// default serde accepts BOTH a JSON string (`"1.5"` — what the SPA
+    /// composer now sends, mirroring the C11 Decimal-as-string wire
+    /// convention used for `exchange_rate`) AND a JSON number (`1` —
+    /// pre-S157 side-stored bodies, CLI callers); it serialises back to a
+    /// string. Both legacy and new side-store rows therefore round-trip.
+    pub quantity: Decimal,
     #[serde(rename = "unitPrice")]
     pub unit_price: i64,
     #[serde(rename = "vatRatePercent")]
